rename, move Beautifier --> CommencementsBeautifier
